Create cluster resources component
