+++ src/components/App/index.js
@@ -1,22 +1,9 @@
 import React from 'react';
 import './App.css';
-import Content from '../ResourceCard/Example';
-
-/* These below are for showing purposes */
-const resources = [
-  {
-    "name": "Nodes",
-    "resourceNumber": 7
-  },
-  {
-    "name": "Secrets",
-    "resourceNumber": 3
-  }
-]
-
+import ClusterResource from '../ResourceCard';
 
 function App() {
-  return <div><Content resources={resources}/></div>
+  return <div><ClusterResource /></div>
 }
 
 export default App;
