+++ src/components/ClusterResources/index.js
@@ -1,0 +1,35 @@
+import React from 'react';
+import ResourceCard from '../ResourceCard';
+
+const resources = [
+    {
+      "name": "Nodes",
+      "resourceNumber": 7
+    },
+    {
+      "name": "Secrets",
+      "resourceNumber": 3
+    },
+    {
+      "name": "Nodes",
+      "resourceNumber": 7
+      },
+      {
+      "name": "Secrets",
+      "resourceNumber": 3
+      },
+      {
+      "name": "Nodes",
+      "resourceNumber": 7
+      },
+      {
+      "name": "Secrets",
+      "resourceNumber": 3
+      }
+  ]
+const ClusterResources = () =>
+    resources.map((resource) =>
+        <ResourceCard title={resource.name} resourceCount={resource.resourceNumber} />
+);
+
+export default ClusterResources
